Added register functionality and fixed bugs

+++ store/models.py
@@ -9,7 +9,7 @@ class Book(models.Model):
     genre = models.CharField(max_length=50)
     description = models.TextField(null=True)
     mrp = models.PositiveIntegerField()
-    rating = models.IntegerField(default=0)
+    rating = models.FloatField(default=0)
     total_rated=models.IntegerField(default=0)
 
     class Meta:
@@ -35,8 +35,8 @@ class BookCopy(models.Model):
 class BookRate(models.Model):
     book_rated = models.ForeignKey(Book, on_delete=models.CASCADE)
     user_name=models.ForeignKey(User, on_delete=models.CASCADE,default=None)
-    user_rating=models.IntegerField(
-        default=0, validators=[MaxValueValidator(10), MinValueValidator(0)])
+    user_rating=models.FloatField(
+        default=0.0, validators=[MaxValueValidator(10), MinValueValidator(0)])
 
     def __str__(self):
         return f'{self.book_rated.title}'
